Separate blazar_library by platform
